@@ -1,9 +1,23 @@
 <template>
-    <div>
-    	hello , this app vue test dev bbbbaaaa!
-    	<router-view></router-view>
-    </div>
+  <div id="app">
+    <img src="./assets/logo.png">
+    <router-view/>
+  </div>
 </template>
-<style>  
-    body{color:red;font-size: 20px}
+
+<script>
+export default {
+  name: 'App'
+}
+</script>
+
+<style>
+#app {
+  font-family: 'Avenir', Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+  margin-top: 60px;
+}
 </style>
